Reorganize new plans and utilities; Rename some items
