+++ src/blop/__init__.py
@@ -1,9 +1,9 @@
 """A bridge between optimization algorithms and Bluesky."""
 
+from .plan_stubs import list_scan_in_run
 from .plans import (
     acquire_baseline,
     default_acquire,
-    default_in_run_acquire,
     optimize,
     optimize_in_run,
     optimize_step,
@@ -19,7 +19,7 @@ __all__ = [
     "__version__",
     "acquire_baseline",
     "default_acquire",
-    "default_in_run_acquire",
+    "list_scan_in_run",
     "optimize",
     "optimize_in_run",
     "optimize_step",
