+++ src/blop/plan_stubs.py
@@ -1,16 +1,29 @@
 """Bluesky plan stubs for optimization."""
 
+import logging
 from collections import defaultdict
 from collections.abc import Hashable, Mapping, MutableMapping, Sequence
 from typing import Any, Literal, cast
 
 import bluesky.plan_stubs as bps
+import bluesky.plans as bp
 import numpy as np
+from bluesky.protocols import Readable
 from bluesky.utils import MsgGenerator, plan
 from numpy.typing import ArrayLike
 
-from .protocols import ID_KEY, Actuator, Optimizer
-from .utils import InferredReadable, Source
+from .protocols import ID_KEY, Actuator, Optimizer, Sensor
+from .utils import (
+    InferredReadable,
+    Source,
+    _drop_run_control_messages,
+    _suggestion_ids,
+    _unpack_for_list_scan,
+    _validate_suggestions,
+    route_suggestions,
+)
+
+logger = logging.getLogger(__name__)
 
 _ACQUISITION_UID_KEY: Literal["acquisition_uid"] = "acquisition_uid"
 _SUGGESTION_IDS_KEY: Literal["suggestion_ids"] = "suggestion_ids"
@@ -29,6 +42,27 @@ def _acquisition_identifier_value(uid: Hashable) -> ArrayLike:
     if _is_array_like_identifier(uid):
         return cast(ArrayLike, uid)
     return repr(uid)
+
+
+@plan
+def seq_read(readables: Sequence[Readable], **kwargs: Any) -> MsgGenerator[dict[str, Any]]:
+    """
+    Read the current values of the given readables.
+
+    Parameters
+    ----------
+    readables : Sequence[Readable]
+        The readables to read.
+
+    Returns
+    -------
+    dict[str, Any]
+        A dictionary of the readable names and their current values.
+    """
+    results = {}
+    for readable in readables:
+        results[readable.name] = yield from bps.rd(readable, **kwargs)
+    return results
 
 
 @plan
@@ -188,3 +222,75 @@ def navigate_to_best(
 
     if moves:
         yield from bps.mv(*moves)
+
+
+@plan
+def list_scan_in_run(
+    suggestions: Sequence[Mapping],
+    actuators: Sequence[Actuator],
+    sensors: Sequence[Sensor] | None = None,
+    *,
+    per_step: bp.PerStep | None = None,
+    **kwargs: Any,
+) -> MsgGenerator[tuple[Hashable, ...]]:
+    """Acquire suggestions inside an already-open Bluesky run.
+
+    This plan moves through the suggestions, optionally reordering them for efficient motion,
+    and executes a Bluesky list scan without opening a child run. The list scan's stage and
+    unstage messages are preserved.
+
+    Parameters
+    ----------
+    suggestions : Sequence[Mapping]
+        Suggested parameterizations to execute. Each suggestion must contain a hashable ``_id``.
+    actuators : Sequence[Actuator]
+        Actuators to move to the suggested positions.
+    sensors : Sequence[Sensor] | None, optional
+        Sensors that produce data to evaluate. Non-readable sensors are ignored.
+    per_step : bp.PerStep | None, optional
+        Bluesky list-scan step hook. Custom hooks may emit any number of events into any streams.
+    **kwargs : Any
+        Additional keyword arguments to pass to :func:`bluesky.plans.list_scan`.
+
+    Returns
+    -------
+    tuple[Hashable, ...]
+        Suggestion IDs in the order the suggestions were executed.
+
+        This identifier intentionally does not encode stream names, event UIDs, event counts, or
+        per-stream offsets. Custom ``per_step`` hooks may emit any number of events into any number
+        of streams. The matching evaluation function is responsible for interpreting those documents
+        and correlating them with these ordered suggestion IDs.
+
+    Yields
+    ------
+    Msg
+        Bluesky messages.
+    """
+    _validate_suggestions(suggestions)
+    if sensors is None:
+        sensors = []
+    readables = [s for s in sensors if isinstance(s, Readable)]
+    if len(readables) != len(sensors):
+        logger.warning(f"Some sensors are not readable and will be ignored. Using only the readable sensors: {readables}")
+
+    if len(suggestions) > 1:
+        if all(isinstance(actuator, Readable) for actuator in actuators):
+            current_position = yield from seq_read(cast(Sequence[Readable], actuators))
+        else:
+            current_position = None
+        suggestions = route_suggestions(suggestions, starting_position=current_position)
+
+    suggestion_ids = _suggestion_ids(suggestions)
+    plan_args = _unpack_for_list_scan(suggestions, actuators)
+    # TODO: fix argument type in bluesky.plans.list_scan
+    yield from _drop_run_control_messages(
+        bp.list_scan(
+            readables,
+            *plan_args,  # type: ignore[arg-type]
+            per_step=per_step,
+            **kwargs,
+        )
+    )
+
+    return suggestion_ids
